updated options and logic of compile.py

+++ compile.py
@@ -22,13 +22,20 @@ def prl():
 ### Perform operation in each repo path ######
 ##############################################
 
-def perform_operations(name,git_versions,special_branches,keep_branches):
+def perform_operations(name,git_versions,special_branches):
+  good = True
   print("In "+name+".")
+  print()
+  print("Absolute path:\n"+os.getcwd())
+  print()
+  print("Remote repository:")
+  os.system("git remote -v")
+  print()
   print("Current branch:")
   os.system("git branch -a")
   print()
-  if not keep_branches and special_branches \
-                       and name in special_branches.keys():
+  if '-kb' not in sys.argv and '-r' not in sys.argv \
+     and special_branches and name in special_branches.keys():
     branch = special_branches[name]
     print("Changing to branch:",branch)
     # print("git checkout "+branch)
@@ -36,8 +43,8 @@ def perform_operations(name,git_versions,special_branches,keep_branches):
     print()
     print("After change:")
     os.system("git branch -a")
-  if '-rh' in sys.argv:
-    if not keep_branches and special_branches \
+  if '-rh' in sys.argv and '-r' not in sys.argv:
+    if '-kb' not in sys.argv and special_branches \
                          and name in special_branches.keys():
       branch = special_branches[name]
     else:
@@ -49,14 +56,29 @@ def perform_operations(name,git_versions,special_branches,keep_branches):
     print("After change:")
     os.system("git branch -a")
   if '-p' in sys.argv and '-r' not in sys.argv:
-    print()
-    print("Pulling from remote origin.")
-    print()
-    #print("git pull")
-    os.system("git pull")
-    print()
+    if 'detached' in subprocess.check_output(['git', 'branch'],
+                                             text=True,
+                                             stderr=subprocess.PIPE):
+      print("WARNING: You are in a 'detached HEAD' state.")
+      print()
+      print("Pulling with a detached HEAD is not healthy for you. :(")
+      print("I can only proceed with pulling after restoring HEAD to")
+      print("either the master branch or branches specified in the")
+      print("special_branches file.")
+      print()
+      op = input("Do you want to proceed with\n"
+                 "pulling after restoring head? (y/n): ")
+      if op != 'y':  good =  False
+      else        :  sys.argv.append('-rh')
+    if good:
+      print()
+      print("Pulling from remote origin.")
+      print()
+      #print("git pull")
+      os.system("git pull")
+      print()
   elif '-r' in sys.argv and '-p' not in sys.argv \
-       and '-rh' not in sys.argv and git_versions:
+       and '-rh' not in sys.argv and git_versions and '-kb' not in sys.argv:
     print("Checking out version:")
     # print(git_versions[name])
     os.system("git show "+git_versions[name]+" | head -3")
@@ -68,7 +90,7 @@ def perform_operations(name,git_versions,special_branches,keep_branches):
     os.system("git status | head -1")
     
   
-  if '-r' not in sys.argv:
+  if '-r' not in sys.argv and good:
     print()
     print("Current git repo version:")
     # os.system("git rev-parse HEAD")
@@ -108,6 +130,7 @@ if '-h' in sys.argv:
   print("-rh                  : reset head to either special branches or master")
   print("-kb                  : keep current branches")
   print("-nc                  : do not compile")
+  print("-ncl                 : do not do 'make clean'")
   print("-h                   : print help info")
   print()
   print("When pulling from remote, the current stable git repo versions will")
@@ -130,11 +153,12 @@ else:
   #  prl()
   #  os.system("make git_pull")
   #  prl()
-    
-  print("Make clean:")
-  prl()
-  os.system("make clean")
-  prl()
+  
+  if '-nc' not in sys.argv and '-ncl' not in sys.argv:
+    print("Make clean:")
+    prl()
+    os.system("make clean")
+    prl()
   
   tokens = (str(datetime.datetime.now())).split()
   tokens[1] = tokens[1].split('.')[0]
@@ -160,26 +184,41 @@ else:
   git_versions = {}
   
   if '-p' in sys.argv and '-r' not in sys.argv:
-    exe = "exe/bam"
-    if not os.path.isfile(exe):
-      exe = "exe/nmesh"
-      if not os.path.isfile(exe):
-        exe = "exe/sgrid"
-        if not os.path.isfile(exe):
-          exe = None
-          print("No current executable")
-          op = input("Do you want to proceed? (y/n): ")
-          if op != 'y':  good =  False
     
-    if exe and '-nc' not in sys.argv:
-      print("Keeping current stable executable as "+exe+"_stable."+cur_time+".")
-      os.system("mv "+exe+" "+exe+"_stable."+cur_time)
+    if 'detached' in subprocess.check_output(['git', 'branch'],
+                                             text=True,
+                                             stderr=subprocess.PIPE):
+      print("WARNING: You are in a 'detached HEAD' state.")
+      print()
+      print("Pulling with a detached HEAD is not healthy for you. :(")
+      print("I can only proceed with pulling after restoring HEAD to")
+      print("either the master branch or branches specified in the")
+      print("special_branches file.")
+      print()
+      op = input("Do you want to proceed with\n"
+                 "pulling after restoring head? (y/n): ")
+      if op != 'y':  good =  False
+      else        :  sys.argv.append('-rh')
+    if good :
+      exe = "exe/bam"
+      if not os.path.isfile(exe):
+        exe = "exe/nmesh"
+        if not os.path.isfile(exe):
+          exe = "exe/sgrid"
+          if not os.path.isfile(exe):
+            exe = None
+            print("No current executable")
+      
+      if exe and '-nc' not in sys.argv:
+        print("Keeping current stable executable as "+exe+"_stable."+cur_time+".")
+        os.system("mv "+exe+" "+exe+"_stable."+cur_time)
     
   #  print("Adding key to agent.")
   #  os.system("eval `add-key.py ~/.ssh/id_rsa`")
   #  print("Renaming git_repo_versions.txt to git_repo_versions_last.txt.")
   #  os.system("mv "+git_ver_file+" "+git_ver_file_last)
-  elif '-r' in sys.argv and '-p' not in sys.argv and '-rh' not in sys.argv:
+  if '-r' in sys.argv and '-p' not in sys.argv \
+    and '-rh' not in sys.argv and '-kb' not in sys.argv:
     try:
       git_ver_file_last = sys.argv[sys.argv.index('-r')+1]
     except:
@@ -190,11 +229,12 @@ else:
         print("Reading git repo versions to revert to\n"
               "from git repo versions file:")
         print(git_ver_file_last)
-        git_versions = {line.split()[0]: line.split()[1] for line in file}
+        git_versions = {line.split()[0]: line.split()[-1] for line in file}
         if not git_versions:
           print("No git repo version information found.")
           op = input("Do you want to proceed? (y/n): ")
           if op != 'y':  good =  False
+        
     else:
       print("Last git repo versions file not found.")
       print("Choosing default configuration: no -p and no -r.")
@@ -212,23 +252,15 @@ else:
     print("Choosing default configuration: no -r and no -rh.")
     op = input("Do you want to proceed? (y/n): ")
     if op != 'y':  good =  False
-  elif '-kb' in sys.argv and '-p' in sys.argv:
-    print("Pulling while keeping existing branches.")
+  elif '-r' in sys.argv and '-kb' in sys.argv:
+    print("Cannot specify -r and -kb options at once.")
+    print("Choosing default configuration: no -r and no -kb.")
     op = input("Do you want to proceed? (y/n): ")
     if op != 'y':  good =  False
-    else        :  keep_branches = True
-  elif '-kb' in sys.argv and '-r' in sys.argv:
-    print("Reverting git repo versions while keeping existing branches.")
-    op = input("Do you want to proceed? (y/n): ")
-    if op != 'y':  good =  False
-    else        :  keep_branches = True
-  elif '-kb' in sys.argv and '-rh' in sys.argv:
-    print("Restoring git repo head while keeping existing branches.")
-    op = input("Do you want to proceed? (y/n): ")
-    if op != 'y':  good =  False
-    else        :  keep_branches = True
-  else:
-    if '-kb' in sys.argv:  keep_branches = True
+  
+  if good:
+    
+    special_branches = {}
     exe = "exe/bam"
     if not os.path.isfile(exe):
       exe = "exe/nmesh"
@@ -240,34 +272,33 @@ else:
     if exe and '-nc' not in sys.argv:
       os.system("mv "+exe+" "+exe+"."+cur_time)
   
-  if '-r' not in sys.argv and good:
-    print("Opening git repo versions file to write into:")
-    print(git_ver_file)
-    f = open(git_ver_file,"w+")
-  
-  if good:
-    prl()
-    special_branches = {}
-    print("Checking for special_branches file.")
-    if os.path.isfile("special_branches"):
-      with open("special_branches") as file:
-        print("Reading special_branches file.")
-        # print("special_branches")
-        special_branches = {line.split()[0]: line.split()[1] for line in file}
-        if not special_branches:
-          print("No special branch information found.")
+    if '-r' not in sys.argv:
+      print("Opening git repo versions file to write into:")
+      print(git_ver_file)
+      f = open(git_ver_file,"w+")
+    
+      prl()
+      if '-kb' not in sys.argv:
+        print("Checking for special_branches file.")
+        if os.path.isfile("special_branches"):
+          with open("special_branches") as file:
+            print("Reading special_branches file.")
+            # print("special_branches")
+            special_branches = {line.split()[0]: line.split()[1] for line in file}
+            if not special_branches:
+              print("No special branch information found.")
+              op = input("Do you want to proceed? (y/n): ")
+              if op != 'y':  good =  False
+        else:
+          print("No special_branches file found.")
+          print("Keeping pre-checked branches for all repos.")
           op = input("Do you want to proceed? (y/n): ")
           if op != 'y':  good =  False
-    else:
-      print("No special_branches file found.")
-      print("Keeping pre-checked branches for all repos.")
-      op = input("Do you want to proceed? (y/n): ")
-      if op != 'y':  good =  False
   
   if good:
   
     prl()
-    perform_operations("main",git_versions,special_branches,keep_branches)
+    perform_operations("main",git_versions,special_branches)
     
     if projects_dir:
     
@@ -281,28 +312,30 @@ else:
       # print(git_versions)
       
       for d in dls_all: 
-        if not d.startswith('.') and os.path.isdir(d) :
+        if not d.startswith('.') and os.path.isdir(d) and good:
           prl()
           os.chdir(d)
-          perform_operations(d,git_versions,special_branches,keep_branches)
+          perform_operations(d,git_versions,special_branches)
           os.chdir("..")
   
     if '-r' not in sys.argv:
       print("Closing git repo versions file.")
       f.close()
     
-    prl()
-    print("Done checking out Branches and/or versions. :)")
-    if '-nc' not in sys.argv:
-      print("Now, building executable. :)")
-      os.chdir(top)
-      print("Make clean:")
+    if good:
       prl()
-      os.system("make clean")
-      prl()
-      prl()
-      #print("make -j 12")
-      os.system("make -j 12")
+      print("Done checking out Branches and/or versions. :)")
+      if '-nc' not in sys.argv:
+        print("Now, building executable. :)")
+        os.chdir(top)
+        if '-ncl' not in sys.argv:
+          print("Make clean:")
+          prl()
+          os.system("make clean")
+          prl()
+        prl()
+        #print("make -j 12")
+        os.system("make -j 12")
     prl()
     print("All done! :)")
